styling changes to expert dashboard header and upcoming appointments
