+++ codebase/components/Experts/ExpertsDashboard/messages.js
@@ -1,0 +1,19 @@
+const messages = [
+    "Welcome, Expert Extraordinaire! Prepare to unleash your mental health superpowers.",
+    "Hey there, Healing Hero! Your expertise is the key to HappyHead's success.",
+    "HappyHead salutes you, Mental Health Maestro! Together, we'll make minds bloom.",
+    "Step into your Zen Zone, Wise Wizard of Wellness. Your guidance is invaluable.",
+    "Greetings, Mind Magician! Get ready to cast spells of healing and positivity.",
+    "Welcome back, Champion of Change! Your expertise lights the path to happiness.",
+    "HappyHead says: Let your expertise shine, Mental Marvel. The world needs your magic.",
+    "Hello, Mind Mastermind! Buckle up for another day of mental health triumphs.",
+    "Step into the HappyHead headquarters, Brainy Trailblazer. Let's revolutionize mental well-being.",
+    "HappyHead: Where experts like you turn compassion into superpowers.",
+    "Welcome, Guardian of Good Vibes! Get ready to sprinkle positivity like confetti.",
+    "HappyHead bows to your brilliance, Mental Health Guru. Lead the way to inner peace.",
+    "Hey there, Empathy Engineer! Your presence makes HappyHead an oasis of support.",
+    "HappyHead says: Today, you'll weave wonders with your words, Expert Enchanter.",
+    "Welcome back, Doctor of Delight! Your knowledge is the remedy for weary souls."
+  ];
+
+  export default messages
